product exception handling dao complete

+++ src/main/java/com/danaga/dao/product/RecentViewDaoImpl.java
@@ -7,8 +7,17 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
 import com.danaga.dto.product.RecentViewDto;
+import com.danaga.entity.Member;
 import com.danaga.entity.OptionSet;
 import com.danaga.entity.RecentView;
+import com.danaga.exception.product.AlreadyExistsException;
+import com.danaga.exception.product.AlreadyExistsException.ExistsRecentViewException;
+import com.danaga.exception.product.FoundNoObjectException;
+import com.danaga.exception.product.FoundNoObjectException.FoundNoMemberException;
+import com.danaga.exception.product.FoundNoObjectException.FoundNoOptionSetException;
+import com.danaga.exception.product.FoundNoObjectException.FoundNoRecentViewException;
+import com.danaga.repository.MemberRepository;
+import com.danaga.repository.product.OptionSetRepository;
 import com.danaga.repository.product.RecentViewRepository;
 
 import jakarta.persistence.EntityManager;
@@ -20,24 +29,40 @@ import lombok.RequiredArgsConstructor;
 public class RecentViewDaoImpl implements RecentViewDao{
 
 	private final RecentViewRepository repository;
+	private final MemberRepository memberRepository;
+	private final OptionSetRepository optionSetRepository;
 	
 	@PersistenceContext
 	private EntityManager em;
 	
 	//나의 최근본 상품 하나 선택 삭제
 	@Override
-	public void delete(RecentView entity) {
-		repository.deleteByMemberIdAndOptionSetId(entity.getMember().getId(),entity.getOptionSet().getId());
+	public void delete(RecentView entity) throws FoundNoRecentViewException {
+		RecentView find = repository.findByMemberIdAndOptionSetId(entity.getMember().getId(), entity.getOptionSet().getId());
+		if(find==null) {
+			throw new FoundNoRecentViewException();
+		}
+		repository.deleteByMemberIdAndOptionSetId(find.getMember().getId(),find.getOptionSet().getId());
 	}
 	//나의 최근본 상품 전체 삭제
 	@Override
-	public void deleteAll(Long memberId) {
+	public void deleteAll(Long memberId) throws FoundNoRecentViewException {
+		List<RecentView> find = repository.findByMemberId(memberId);
+		if(find==null) {
+			throw new FoundNoRecentViewException();
+		}
 		repository.deleteByMemberId(memberId);
 	}
 
 	//recentView 추가
 	@Override
-	public RecentView save(RecentView entity) {
+	public RecentView save(RecentView entity) throws FoundNoMemberException, FoundNoOptionSetException, ExistsRecentViewException {
+		Member findMember = memberRepository.findById(entity.getMember().getId()).orElseThrow(() -> new FoundNoMemberException());
+		OptionSet findOptionSet = optionSetRepository.findById(entity.getOptionSet().getId()).orElseThrow(()-> new FoundNoOptionSetException());
+		RecentView find=repository.findByMemberIdAndOptionSetId(entity.getMember().getId(), entity.getOptionSet().getId());
+		if(find!=null) {
+			throw new ExistsRecentViewException();
+		}
 		return repository.save(entity);
 	}
